refactor: fetchJson も lib/market-api.ts に共通化し lib/ の重複も解消
レビューで判明した追加の重複を修正：
- jpx-market-closed.ts / us-market-closed.ts にも同一の getApiBaseUrl / fetchJson が存在していた
- fetchJson を lib/market-api.ts に追加し、7ファイル全ての重複定義を削除
