+++ app/tools/earnings-calendar/data-loader.ts
@@ -1,7 +1,7 @@
 import { readFile } from "node:fs/promises";
 import path from "node:path";
 import { loadUsMarketClosedData } from "@/lib/us-market-closed";
-import { getApiBaseUrl } from "@/lib/market-api";
+import { getApiBaseUrl, fetchJson } from "@/lib/market-api";
 import type {
   EarningsCalendarManifest,
   EarningsCalendarItem,
@@ -18,25 +18,6 @@ function getDataDir() {
   return path.join(process.cwd(), "app/tools/earnings-calendar/data");
 }
 
-async function fetchJson<T>(url: string): Promise<T> {
-  const controller = new AbortController();
-  const timeoutId = setTimeout(() => controller.abort(), 5000);
-
-  try {
-    const res = await fetch(url, {
-      signal: controller.signal,
-      next: { revalidate: 300 },
-    });
-
-    if (!res.ok) {
-      throw new Error(`Failed to fetch ${url}: HTTP ${res.status}`);
-    }
-
-    return (await res.json()) as T;
-  } finally {
-    clearTimeout(timeoutId);
-  }
-}
 
 async function loadLocalDomesticManifest(): Promise<EarningsCalendarManifest> {
   const raw = await readFile(path.join(getDataDir(), "manifest.json"), "utf-8");
