+++ app/tools/topix33/data-loader.ts
@@ -1,6 +1,6 @@
 import { readFile } from "node:fs/promises";
 import path from "node:path";
-import { getApiBaseUrl } from "@/lib/market-api";
+import { getApiBaseUrl, fetchJson } from "@/lib/market-api";
 import type { Topix33DayData, Topix33Manifest } from "./types";
 
 const EMPTY_MANIFEST: Topix33Manifest = {
@@ -10,26 +10,6 @@ const EMPTY_MANIFEST: Topix33Manifest = {
 
 function getDataDir() {
   return path.join(process.cwd(), "app/tools/topix33/data");
-}
-
-async function fetchJson<T>(url: string): Promise<T> {
-  const controller = new AbortController();
-  const timeoutId = setTimeout(() => controller.abort(), 5000);
-
-  try {
-    const res = await fetch(url, {
-      signal: controller.signal,
-      next: { revalidate: 300 },
-    });
-
-    if (!res.ok) {
-      throw new Error(`Failed to fetch ${url}: HTTP ${res.status}`);
-    }
-
-    return (await res.json()) as T;
-  } finally {
-    clearTimeout(timeoutId);
-  }
 }
 
 async function loadLocalManifest(): Promise<Topix33Manifest> {
