+++ app/tools/yutai-candidates/data-loader.ts
@@ -1,6 +1,6 @@
 import { readFile } from "node:fs/promises";
 import path from "node:path";
-import { getApiBaseUrl } from "@/lib/market-api";
+import { getApiBaseUrl, fetchJson } from "@/lib/market-api";
 import type {
   MonthlyYutaiManifest,
   MonthlyYutaiMonthData,
@@ -99,26 +99,6 @@ function getSmartDefaultMonthId(availableMonths: string[], fallback: string): st
 
 function getDataDir() {
   return path.join(process.cwd(), "app/tools/yutai-candidates/data");
-}
-
-async function fetchJson<T>(url: string): Promise<T> {
-  const controller = new AbortController();
-  const timeoutId = setTimeout(() => controller.abort(), 5000);
-
-  try {
-    const res = await fetch(url, {
-      signal: controller.signal,
-      next: { revalidate: 300 },
-    });
-
-    if (!res.ok) {
-      throw new Error(`Failed to fetch ${url}: HTTP ${res.status}`);
-    }
-
-    return (await res.json()) as T;
-  } finally {
-    clearTimeout(timeoutId);
-  }
 }
 
 async function loadLocalManifest(): Promise<MonthlyYutaiManifest | null> {
